refactor(utils): split path identity from path wrappers
